+++ Application/Source/FPSCam.cpp
@@ -51,8 +51,8 @@ void FPSCam::Update(double dt, double dx, double dy)
 
 	float horizontalDISAngle = 0;
 	float verticalDISAngle = 0;
-	horizontalDISAngle += mouseSpeed * dx; //taking displacement as pi
-	verticalDISAngle += mouseSpeed * dy;
+	horizontalDISAngle += mouseSpeed * (float)dx; //taking displacement as pi
+	verticalDISAngle += mouseSpeed * (float)dy;
 
 
 	if (Cam_Theta + Math::RadianToDegree(horizontalDISAngle) > 360.0f)
